Refactoring folder names for the API module
