+++ .gitignore
@@ -76,6 +76,6 @@ covalent_ui/dask-worker-space/*.*lock
 covalent_ui/app/dask-worker-space/*.*lock
 
 # Ignore .sonarwork files
-covalent_ui/os_api/.scannerwork/.sonar_lock
-covalent_ui/os_api/.scannerwork/report-task.txt
-covalent_ui/os_api/sonar.sh
+covalent_ui/api/.scannerwork/.sonar_lock
+covalent_ui/api/.scannerwork/report-task.txt
+covalent_ui/api/sonar.sh
