+++ covalent_ui/api/v1/data_layer/electron_dal.py
@@ -22,10 +22,10 @@ from datetime import datetime, timezone
 
 from sqlalchemy.sql import func
 
-from covalent_ui.os_api.api_v0.database.schema.electrons_schema import Electron
-from covalent_ui.os_api.api_v0.database.schema.lattices_schema import Lattice
-from covalent_ui.os_api.api_v0.utils.file_handle import FileHandler
-from covalent_ui.os_api.api_v0.utils.file_name import FileName
+from covalent_ui.api.v1.database.schema.electrons_schema import Electron
+from covalent_ui.api.v1.database.schema.lattices_schema import Lattice
+from covalent_ui.api.v1.utils.file_handle import FileHandler
+from covalent_ui.api.v1.utils.file_name import FileName
 
 
 class Electrons:
