+++ covalent_ui/api/v1/data_layer/graph_dal.py
@@ -24,10 +24,10 @@ from uuid import UUID
 from fastapi import HTTPException
 from sqlalchemy.orm import Session
 
-from covalent_ui.os_api.api_v0.database.schema.electron_dependency import ElectronDependency
-from covalent_ui.os_api.api_v0.database.schema.electrons_schema import Electron
-from covalent_ui.os_api.api_v0.database.schema.lattices_schema import Lattice
-from covalent_ui.os_api.api_v0.models.lattices_model import GraphResponse
+from covalent_ui.api.v1.database.schema.electron_dependency import ElectronDependency
+from covalent_ui.api.v1.database.schema.electrons_schema import Electron
+from covalent_ui.api.v1.database.schema.lattices_schema import Lattice
+from covalent_ui.api.v1.models.lattices_model import GraphResponse
 
 
 class Graph:
