+++ covalent_ui/api/v1/data_layer/summary_dal.py
@@ -26,9 +26,9 @@ from sqlalchemy import update
 from sqlalchemy.orm import Session
 from sqlalchemy.sql import desc, func, or_
 
-from covalent_ui.os_api.api_v0.database.schema.electrons_schema import Electron
-from covalent_ui.os_api.api_v0.database.schema.lattices_schema import Lattice
-from covalent_ui.os_api.api_v0.models.dispatch_model import (
+from covalent_ui.api.v1.database.schema.electrons_schema import Electron
+from covalent_ui.api.v1.database.schema.lattices_schema import Lattice
+from covalent_ui.api.v1.models.dispatch_model import (
     DeleteDispatchesRequest,
     DeleteDispatchesResponse,
     DispatchDashBoardResponse,
