+++ covalent_ui/api/v1/database/schema/electrons_schema.py
@@ -18,22 +18,47 @@
 #
 # Relief from the License may be granted by purchasing a commercial license.
 
-"""Lattices schema"""
-from sqlalchemy import BigInteger, Boolean, Column, DateTime, Integer, String, Text
-from sqlalchemy.orm import relationship
+"""Electrons Schema """
+import enum
 
-from covalent_ui.os_api.api_v0.database.config.db import Base
+from sqlalchemy import BigInteger, Boolean, Column, DateTime, ForeignKey, Integer, String, Text
+
+from covalent_ui.api.v1.database.config.db import Base
 
 
-class Lattice(Base):
-    """Lattice
+class ElectronTypeEnum(enum.Enum):
+    """Electron Type Of Enum
+
+    Attributes:
+        electron
+        parameter
+        sublattice
+        collection
+        attribute
+        generated
+        subscript
+    """
+
+    electron = 1
+    parameter = 2
+    sublattice = 3
+    collection = 4
+    attribute = 5
+    generated = 6
+    subscript = 7
+    function = 8
+    electron_list = 9
+
+
+class Electron(Base):
+    """Electron
 
     Attributes:
         id: primary key id
-        dispatch_id: id of the dispatch
-        electron_id: id of node if the lattice is actually a sublattice
-        name: name of the lattice function
-        status: workflow status
+        parent_lattice_id: id of the lattice containing this electron
+        transport_graph_node_id: id of the node in the context of a transport graph
+        type: node type
+        name: node name
         status: Execution status of the node
         storage_type: Storage backend type for data files ("local", "s3")
         storage_path: Bucket name (dispatch_id)
@@ -42,27 +67,30 @@ class Lattice(Base):
         executor_filename: Name of the file containing the serialized executor
         error_filename: Name of the file containing an error message for the electron
         results_filename: Name of the file containing the serialized output
-        inputs_filename: Name of the file containing the serialized input data
+        value_filename: Name of the file containing the serialized parameter (for parameter nodes)
+        attribute_name: Electron attribute name (for attribute nodes)
+        key: For generated and subscript nodes
+        stdout_filename: Name of the file containing standard output generated by the task
+        stderr_filename: Name of the file containing standard error generated by the task
+        info_filename: Name of the file containing execution information generated at runtime
         created_at: created timestamp
         updated_at: updated timestamp
         started_at: started timestamp
         completed_at: completed timestamp
     """
 
-    __tablename__ = "lattices"
+    __tablename__ = "electrons"
     id = Column(BigInteger, primary_key=True)
-    dispatch_id = Column(String(64), nullable=False)
 
-    electron_id = Column(Integer)
+    parent_lattice_id = Column(BigInteger, ForeignKey("lattices.id"), nullable=False)
+
+    transport_graph_node_id = Column(Integer, nullable=False)
+
+    type = Column(String(24), nullable=False)
 
     name = Column(Text, nullable=False)
 
     status = Column(String(24), nullable=False)
-
-    electron_num = Column(Integer, nullable=False)
-
-    # Number of completed nodes in the lattice
-    completed_electron_num = Column(Integer, nullable=False)
 
     storage_type = Column(Text)
 
@@ -74,12 +102,25 @@ class Lattice(Base):
 
     executor_filename = Column(Text)
 
-    error_filename = Column(Text)
-
-    inputs_filename = Column(Text)
-
     results_filename = Column(Text)
-    transport_graph_filename = Column(Text)
+
+    value_filename = Column(Text)
+
+    attribute_name = Column(Text)
+
+    key_filename = Column(Integer)
+
+    stdout_filename = Column(Text)
+
+    deps_filename = Column(Text)
+
+    call_before_filename = Column(Text)
+
+    call_after_filename = Column(Text)
+
+    stderr_filename = Column(Text)
+
+    info_filename = Column(Text)
 
     is_active = Column(Boolean)
